cohesive first paragraph
The first paragraph introduces the full complexity of TensorFlow.
I think it's a mistake to split it into two paragraphs. The
second one in particular ("The execution model...") doesn't stand
alone very well.

The first sentence needs to be a thesis for the paragraph. The
intent is to draw in both people who think they already know
something about TensorFlow and people who don't. The target
audience already knows TensorFlow has buzz around it, so I don't
want to lead with that.

It isn't true that "TensorFlow has been dominating the trending
page on [GitHub] since Google open sourced it". This is the
trending page on GitHub:

https://github.com/trending

TensorFlow isn't there today, and I don't know the last time it
was, or if anyone quantified how many days it was there. The
machine learning "showcase",

https://github.com/showcases/machine-learning

was linked to me by John as well. I think there are a couple
reasons not to include it:

 * The first text people might read there is "Laying the
   foundations for Skynet". It's cute if you get the reference
   and agree that it's funny, but I don't want the distraction.

 * It isn't obvious what that page _is_. You have to figure out
   that it's a curated list of repos, and then see that it's
   sorted in decreasing order of number of stars, to even get the
   point that TensorFlow has a lot of stars. And who cares about
   stars? Not everyone.

If you want to see which repos have the most stars overall, you
can:

https://github.com/search?utf8=%E2%9C%93&q=stars%3A%3E1&type=Repositories&ref=searchresults

TensorFlow was on page four of those results when I looked.

The "Yet Another X" trope is endearing to some part of the
potential audience, but not everyone will be familiar with it.

The word "actually" should be avoided in speech and writing.

If we take out the "Especially" in "Especially for someone hoping
to explore machine learning for the first time" it makes it seem
like the article is _only_ for absolute beginners, which I don't
think is true.