don't show `id()`
I don't want to show `id()` calls and output. It's numbers that
we don't need to bother the reader with, and it's a distraction
from where we're trying to go.

I have some thought of dropping the whole intro Python section,
but I think it has some utility in introducing things, and it
might provide a familiar entry point for Python people.