clarify execution / action
Focusing on "the `append` method is executed right away" moves
farther from what I think this paragraph should convey. With
either a standard Python execution model or the TensorFlow graph
construction model, Python methods are always executed right
away. The difference is that with TensorFlow the method execution
changes the graph, which is separate from what we usually think
of as evaluation.

I removed the code font and changed the wording a little bit to
hopefully make it clearer when I'm talking about `append` the
method and "append" the action.

I probably use it some myself, but it's good to avoid passive voice.