more different links in closing
I don't want to use the technique of linking subsequent words to
related but not specifically referenced things. I also don't
think it's a natural transition to jump to images without more
introduction. A beginner would struggle a lot to do anything on
their own with CIFAR, for example. I think it is a good idea to
link back again to the tutorials and how-tos though - thanks!